v0.1.2
现在编译时遇到未定义函数不再报错，而是等待后续定义。
添加一种触发器试试效果

+++ class/ScriptSuperPointer.cpp
@@ -55,7 +55,7 @@ namespace zlscript
 		}
 		return false;
 	}
-	int CScriptPointInterface::GetScriptPointIndex()
+	__int64 CScriptPointInterface::GetScriptPointIndex()
 	{
 		if (!IsInitScriptPointIndex())
 		{
